moved packages into another dirs
