+++ backend/usecase/userUsecase.go
@@ -43,30 +43,3 @@ func (uc *UserUsecase) LogoutUser(userId int) error {
 func (uc *UserUsecase) UpdateProfile(userId int, userProfile *models.UserProfile) error {
 	return uc.repo.UpdateProfile(userId, userProfile)
 }
-
-func (uc *UserUsecase) SaveProfilePhoto() error {
-	// minioClient, err :=
-	// if err != nil {
-	// 	return "", err
-	// }
-
-	// objectName := fileHeader.Filename // Можно добавить уникальность
-	// contentType := fileHeader.Header.Get("Content-Type")
-
-	// // Загрузка файла в MinIO
-	// info, err := minioClient.PutObject(
-	// 	context.Background(),
-	// 	bucketName,
-	// 	objectName,
-	// 	file,
-	// 	fileHeader.Size,
-	// 	minio.PutObjectOptions{ContentType: contentType},
-	// )
-	// if err != nil {
-	// 	return "", err
-	// }
-
-	// fileURL := fmt.Sprintf("https://%s/%s/%s", endpoint, bucketName, info.Key)
-	// return fileURL, nil
-	return nil
-}
